# clients/openapi.yaml
openapi: 3.1.0
info:
  title: mem-dog API
  version: 4.0.0
  description: >
    Private AI system for multi-channel data ingestion, AI enrichment,
    temporal knowledge graph, and multi-signal semantic search.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key (md_*), JWT, or global API key
  schemas:
...
    OrgRole:
      type: string
      enum: [owner, admin, member, viewer]
...
    Organization:
      type: object
      properties:
        org_id:
          type: string
          example: org_01HXYZ
        name:
          type: string
        description:
          type: string
...
paths:
...
  /api/v1/list:
    get:
      tags: [List]
      summary: List user data
      operationId: listUserData
      parameters:
        - name: user
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [meta, raw]
            default: meta
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: User data
  /api/v1/list/{data_id}:
    get:
      tags: [List]
      summary: Get single data item
      operationId: listUserDataItem
      parameters:
        - name: data_id
          in: path
          required: true
          schema:
            type: string
        - name: user
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [meta, raw]
      responses:
        '200':
          description: Data item

  # ---------- Tags ----------
...
  /api/v1/organizations:
    post:
      tags: [Organizations]
      summary: Create organization
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
              required: [name]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
    get:
      tags: [Organizations]
      summary: List caller's organizations
      operationId: listOrganizations
      responses:
        '200':
          description: Organizations
  /api/v1/organizations/{org_id}:
    get:
      tags: [Organizations]
      summary: Get organization
      operationId: getOrganization
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Organization
    put:
      tags: [Organizations]
      summary: Update organization
      operationId: updateOrganization
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
    delete:
      tags: [Organizations]
      summary: Delete organization
      operationId: deleteOrganization
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /api/v1/organizations/{org_id}/members:
    post:
      tags: [Organizations]
      summary: Add member
      operationId: addOrgMember
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                role:
                  $ref: '#/components/schemas/OrgRole'
              required: [user_id]
      responses:
        '200':
          description: Added
    get:
      tags: [Organizations]
      summary: List members
      operationId: listOrgMembers
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Members
  /api/v1/organizations/{org_id}/members/{user_id}:
    put:
      tags: [Organizations]
      summary: Update member role
      operationId: updateOrgMember
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  $ref: '#/components/schemas/OrgRole'
              required: [role]
      responses:
        '200':
          description: Updated
    delete:
      tags: [Organizations]
      summary: Remove member
      operationId: removeOrgMember
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Removed

  # ---------- Projects ----------
...